Move WeightInfo, Config, Event into Group pallet.
