+++ pallets/group/src/benchmarking.rs
@@ -1,6 +1,5 @@
 use crate::*;
 use pallet_identity::benchmarking::{User, UserBuilder};
-use polymesh_common_utilities::traits::group::Config;
 use polymesh_primitives::traits::group::GroupTrait;
 
 use frame_benchmarking::benchmarks_instance;
